validator-poc/requirements.txt
pyyaml==6.0.1
requests==2.32.4
pytest==7.4.4
pytest-mock==3.12.0
pytest-cov==4.1.0